work/java: change int i to long i

+++ medusa_software/work/java/org/medusa_re/medusa/main.java
@@ -25,7 +25,7 @@ public class main {
 
 		long total = 0;
 
-		for (int i = 0; i < 0x1000000; i++) {
+		for (long i = 0; i < 0x1000000; i++) {
 			total += i;
 		}
